users-service/requirements.txt
Flask==3.1.0
boto3==1.26.49
flasgger==0.9.7.1
gunicorn==23.0.0
gevent==24.11.1
psycopg2-binary
Flask-SQLAlchemy
requests
pyjwt